requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
stripe==13.2.0
typing_extensions==4.15.0
urllib3==2.5.0
